Setup CFP Page with new language system
